third_party/blink/web_tests/external/wpt/html/semantics/forms/the-output-element/mutations.window-expected.txt

This is a testharness.js-based test.
FAIL Descendant mutations and output.value and .defaultValue assert_equals: .defaultValue after setting textContent expected "x" but got ""
PASS output and output.form.reset()
Harness: the test ran to completion.
